feat(ShellBox): fix unit test
